Read the file to be uploaded
Begin the upload script by testing to confirm we get filename and some
data back from fs.readFile.

Use an empty string as a default argument for the readFile so if
we don't pass an argument when running the script,
we get a sensible error message.

Add line in package.json to run script from the command line.